Disable makehelp target
Because emscripten builds makehelp as a WASM app, and because WASM apps
cannot be run like "normal" apps and have to be launched via Node, this
target is broken.

Ideally we should probably force this to compile with GCC even when
using Emscripten, since it's really part of the build and not part of
the binary distribution.